app/requirements.txt
Flask==2.3.3
gunicorn==21.2.0
boto3==1.38.26
python-dotenv==1.0.1